♻️ extract validated onboarding projection
