+++ libs/frontend/state/conversation/workspace/adapter/README.md
@@ -24,7 +24,7 @@ failures to fixed, display-safe categories.
   read-only result instead of pretending that it is a direct, group, or Agent-session conversation.
 
 DTO mappers and their narrowed conversation DTO shapes are package-private. The onboarding read reuses the
-onboarding state package's model-adjacent validator before reducing the valid snapshot to history. Consumers
+user-onboarding model package's validator before reducing the valid snapshot to history. Consumers
 import the gateway only from the package barrel.
 
 ## Boundary
@@ -39,7 +39,7 @@ transcript. The adapter requests archived conversation rows so the feature can k
 ## Dependency direction
 
 The package carries `scope:conversation-workspace` and `frontend-role:adapter`. It depends inward on
-frontend core, conversation models, the onboarding projection validator, and the workspace state port. It
+frontend core, conversation and user-onboarding models, and the workspace state port. It
 must not import a feature, element, backend package, app, or concrete conversation stream.
 
 ## See also
