+++ libs/frontend/state/onboarding/projection/project.json
@@ -1,0 +1,10 @@
+{
+  "name": "frontend-state-onboarding-projection",
+  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "sourceRoot": "libs/frontend/state/onboarding/projection/src",
+  "tags": ["type:lib", "layer:frontend", "scope:persona-onboarding", "frontend-role:state"],
+  "targets": {
+    "lint": { "executor": "nx:run-commands", "options": { "command": "tsc -p tsconfig.json --noEmit", "cwd": "libs/frontend/state/onboarding/projection" } }
+  }
+}
