+++ libs/frontend/state/onboarding/projection/src/index.ts
@@ -1,0 +1,3 @@
+/** Frontend projection vocabulary; validation stays in the dependency-bottom model package. */
+export { PersonaFirstChatArchetypes, PersonaFirstChatColours, PersonaFirstChatTranscriptKinds, PersonaFirstChatTranscriptRoles, UserOnboardingRouteStates } from "@opencrane/models/user-onboarding";
+export type { PersonaFirstChatContentRevision, PersonaFirstChatCurrentQuestion, PersonaFirstChatPersona, PersonaFirstChatSnapshot, PersonaFirstChatTranscriptEntry } from "@opencrane/models/user-onboarding";
