+++ libs/frontend/state/onboarding/projection/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../../../../tsconfig.json",
+  "compilerOptions": { "noEmit": true },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist"]
+}
