+++ libs/models/user-onboarding/main/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./persona-first-chat.types.js";
+export { ___ParsePersonaFirstChatSnapshot } from "./persona-first-chat.validator.js";
