+++ libs/models/user-onboarding/main/vitest.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "vitest/config";
+import tsconfigPaths from "vite-tsconfig-paths";
+import { _PackageCacheDir } from "../../../../vitest.cache.js";
+
+/** Vitest configuration for the pure onboarding projection model. */
+export default defineConfig({
+	cacheDir: _PackageCacheDir(import.meta.url),
+	plugins: [tsconfigPaths({ projects: ["../../../../tsconfig.vitest.json"] })],
+	test: { passWithNoTests: false },
+});
